LPS-172468 Create AggregationFilterContainer component
