+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Layout/InfoScreen/InfoScreen.tsx
@@ -16,8 +16,8 @@ import ClayForm, {ClayCheckbox} from '@clayui/form';
 import {
 	Card,
 	Input,
-	getLocalizableLabel,
 	REQUIRED_MSG,
+	getLocalizableLabel,
 	invalidateRequired,
 } from '@liferay/object-js-components-web';
 import React from 'react';
