+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Layout/LayoutScreen/ModalAddObjectLayoutTab.tsx
@@ -22,8 +22,8 @@ import {
 	AutoComplete,
 	FormError,
 	Input,
-	getLocalizableLabel,
 	REQUIRED_MSG,
+	getLocalizableLabel,
 	stringIncludesQuery,
 	useForm,
 } from '@liferay/object-js-components-web';
